Replaces Pubnub by Socket.io
