+++ app/build.gradle
@@ -124,7 +124,7 @@ dependencies {
     compile 'joda-time:joda-time:2.7'
     compile 'com.orhanobut:dialogplus:1.4@aar'
     compile 'com.github.castorflex.smoothprogressbar:library-circular:1.1.0'
-    compile 'com.pubnub:pubnub-android:3.7.2'
+    compile 'com.github.nkzawa:socket.io-client:0.4.2'
 }
 
 // Using lambda functions
