+++ app/src/main/res/values/strings_socket_config.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="socket_address">http://192.168.50.225:3000/</string>
+    <string name="socket_event">chat message</string>
+</resources>
